ATS_Resume_Analyzer/requirements.txt
streamlit
streamlit-extras
PyPDF2
google-genai
